reverting some of brain.py

+++ ml-agents-envs/mlagents/envs/tests/test_envs.py
@@ -55,10 +55,15 @@ def test_step(mock_communicator, mock_launcher):
     env.step()
     batched_step_result = env.get_step_result("RealFakeBrain")
     n_agents = batched_step_result.n_agents()
-    env.set_actions("RealFakeBrain", np.zeros((n_agents, spec.action_shape)))
+    env.set_actions(
+        "RealFakeBrain", np.zeros((n_agents, spec.action_shape), dtype=np.float32)
+    )
     env.step()
     with pytest.raises(UnityActionException):
-        env.set_actions("RealFakeBrain", np.zeros((n_agents - 1, spec.action_shape)))
+        env.set_actions(
+            "RealFakeBrain",
+            np.zeros((n_agents - 1, spec.action_shape), dtype=np.float32),
+        )
     batched_step_result = env.get_step_result("RealFakeBrain")
     n_agents = batched_step_result.n_agents()
     env.set_actions("RealFakeBrain", -1 * np.ones((n_agents, spec.action_shape)))
